Copy the collision update callback in the copy constructor of PlanningScene
The collision update callback (current_world_object_update_callback_) is
copied in the copy constructor of PlanningScene to ensure that the child
planning scene behaves the same way as its parent.
It is reset in `decoupleParent()` to avoid calling the callback on a
cloned planning scene, which would bring confusion.

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>